Move withStyles to components directory
